+++ packages/ui/src/components/ui/skeleton.tsx
@@ -1,15 +1,12 @@
-import type { HTMLAttributes } from 'react';
+import type * as React from 'react';
 import classy from '../../primitives/classy';
 
-export interface SkeletonProps extends HTMLAttributes<HTMLDivElement> {}
+export interface SkeletonProps extends React.HTMLAttributes<HTMLDivElement> {}
 
-export function Skeleton({ className, ...props }: SkeletonProps): JSX.Element {
+export function Skeleton({ className, ...props }: SkeletonProps) {
   return (
     <div
-      className={classy(
-        'bg-muted rounded-md animate-pulse motion-reduce:animate-none',
-        className,
-      )}
+      className={classy('bg-muted rounded-md animate-pulse motion-reduce:animate-none', className)}
       {...props}
     />
   );
