+++ packages/ui/test/components/skeleton.a11y.tsx
@@ -1,6 +1,7 @@
 import { render } from '@testing-library/react';
+import { describe, expect, it } from 'vitest';
 import { axe } from 'vitest-axe';
-import { Skeleton } from './skeleton';
+import { Skeleton } from '../../src/components/ui/skeleton';
 
 describe('Skeleton - Accessibility', () => {
   it('has no accessibility violations', async () => {
@@ -11,10 +12,10 @@ describe('Skeleton - Accessibility', () => {
 
   it('has no violations when used as loading placeholder', async () => {
     const { container } = render(
-      <div role="status">
+      <output>
         <Skeleton className="h-4 w-48" />
         <Skeleton className="h-4 w-32 mt-2" />
-      </div>
+      </output>,
     );
     const results = await axe(container);
     expect(results).toHaveNoViolations();
